# src/main/resources/static/swagger.yaml
openapi: 3.0.3
info:
  title: Springular REST API
  description: Springular REST API
  contact:
    name: Josue Lubaki
    url: https://josue-lubaki.ca
    email: josue.lubaki@uqtr.ca
  license:
    name: Apache License Version 2.0
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Base Url
...
paths:
...
  /auth/register:
    post:
      tags:
        - Registration of accounts
      summary: Register
      description: Register to the application
      operationId: registerUserUsingPOST
      parameters:
        - name: apartment
          in: query
          required: false
          style: form
          schema:
            type: string
        - name: city
          in: query
          required: false
          style: form
          schema:
            type: string
        - name: dob
          in: query
          required: false
          style: form
          schema:
            type: string
            format: date
        - name: email
          in: query
          required: false
          style: form
          schema:
            type: string
        - name: fname
          in: query
          required: false
          style: form
          schema:
            type: string
        - name: imageURL
          in: query
          required: false
          style: form
          schema:
            type: string
            format: binary
        - name: lname
          in: query
          required: false
          style: form
          schema:
            type: string
        - name: password
          in: query
          required: false
          style: form
          schema:
            type: string
        - name: phone
          in: query
          required: false
          style: form
          schema:
            type: string
        - name: role
          in: query
          required: false
          style: form
          schema:
            type: string
        - name: state
          in: query
          required: false
          style: form
          schema:
            type: string
        - name: street
          in: query
          required: false
          style: form
          schema:
            type: string
        - name: zip
          in: query
          required: false
          style: form
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '201':
          description: Successfully created an account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '400':
          description: Bad Request
...
  /api/v1/haircuts/{id}:
    get:
      tags:
        - Haircut consultation
      summary: Get haircut by id
      operationId: getHaircutUsingGET
      parameters:
        - name: id
          in: path
          description: Identifier of the hairstyle to consult
          required: true
          style: simple
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved haircut
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HaircutDTO'
  /management/api/v1/haircuts:
    get:
      tags:
        - Haircut Management Controller
      summary: Get all haircuts
      operationId: getHaircutsUsingGET_1
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HaircutDTO'
        '400':
          description: Bad Request
      security:
        - Authorization:
            - global
    post:
      tags:
        - Haircut Management Controller
      summary: Create haircut
      operationId: createHaircutUsingPOST
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HaircutDTO'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/HaircutDTO'
        '201':
          description: Created
        '400':
          description: Bad Request
      security:
        - Authorization:
            - global
  /management/api/v1/haircuts/{id}:
    put:
      tags:
        - Haircut Management Controller
      summary: Update haircut
      operationId: updateHaircutUsingPUT
      parameters:
        - name: id
          in: path
          description: id
          required: true
          style: simple
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HaircutDTO'
      responses:
        '200':
          description: Updated hair
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/HaircutDTO'
        '400':
          description: Bad Request
      security:
        - Authorization:
            - global
    delete:
      tags:
        - Haircut Management Controller
      summary: Delete haircut
      operationId: deleteHaircutUsingDELETE
      parameters:
        - name: id
          in: path
          description: id
          required: true
          style: simple
          schema:
            type: string
      responses:
        '200':
          description: Deleted hair
        '400':
          description: Bad Request
      security:
        - Authorization:
            - global
    get:
      tags:
        - Haircut Management Controller
      summary: Get haircut by ID
      operationId: getHaircutUsingGET_1
      parameters:
        - name: id
          in: path
          description: id
          required: true
          style: simple
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HaircutDTO'
        '400':
          description: Bad Request
      security:
        - Authorization:
            - global
...
  /api/v1/users/{id}:
    get:
      tags:
        - User Controller
      summary: Get a user by ID
      operationId: getUserUsingGET
      parameters:
        - name: id
          in: path
          description: User ID to consult
          required: true
          style: simple
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved a user
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDTORes'
      security:
        - Authorization:
            - global
    put:
      tags:
        - User Controller
      summary: Update a user by ID
      operationId: updateUserUsingPUT
      parameters:
        - name: id
          in: path
          description: User ID to update
          required: true
          style: simple
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTOReq'
      responses:
        '200':
          description: Successfully updated a user
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDTORes'
      security:
        - Authorization:
            - global
    delete:
      tags:
        - User Controller
      summary: Delete a user by ID
      operationId: deleteMyAccountUsingDELETE
      parameters:
        - name: id
          in: path
          description: User ID to delete
          required: true
          style: simple
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
        '204':
          description: Successfully Deleted a user
      security:
        - Authorization:
            - global
  /management/api/v1/users:
    get:
      tags:
        - User Management Controller
      summary: Get all users
      operationId: getUsersUsingGET
      responses:
        '200':
          description: Successfully retrieved all users
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDTORes'
      security:
        - Authorization:
            - global
    post:
      tags:
        - User Management Controller
      summary: Create a new user
      operationId: registerNewUserUsingPOST
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupDTO'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDTORes'
        '400':
          description: Bad Request
      security:
        - Authorization:
            - global
  /management/api/v1/users/{id}:
    get:
      tags:
        - User Management Controller
      summary: Get a user by ID
      operationId: getUserUsingGET_1
      parameters:
        - name: id
          in: path
          description: user ID to consult
          required: true
          style: simple
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved a user
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDTORes'
      security:
        - Authorization:
            - global
    put:
      tags:
        - User Management Controller
      summary: Update a user by ID
      operationId: updateUserUsingPUT_1
      parameters:
        - name: id
          in: path
          description: user ID to update
          required: true
          style: simple
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTOReq'
      responses:
        '200':
          description: Successfully updated a user
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDTORes'
      security:
        - Authorization:
            - global
    delete:
      tags:
        - User Management Controller
      summary: Delete a user by ID
      operationId: deleteUserUsingDELETE
      parameters:
        - name: id
          in: path
          description: user ID to delete
          required: true
          style: simple
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
        '204':
          description: Successfully Deleted a user
      security:
        - Authorization:
            - global
components:
  schemas:
    AddressDTO:
      title: AddressDTO
      type: object
      properties:
        apartement:
          type: string
        city:
          type: string
        id:
          type: integer
          format: int64
        state:
          type: string
        street:
          type: string
        zip:
          type: string
...
    HaircutDTO:
      title: HaircutDTO
      type: object
      properties:
        description:
          type: string
        estimatedTime:
          type: string
        id:
          type: string
        imageURL:
          type: string
        price:
          type: integer
          format: int32
        title:
          type: string
...
    SignupDTO:
      title: SignupDTO
      type: object
      properties:
        apartment:
          type: string
        city:
          type: string
        dob:
          type: string
          format: date
        email:
          type: string
        fname:
          type: string
        imageURL:
          type: string
        lname:
          type: string
        password:
          type: string
        phone:
          type: string
        role:
          type: string
        state:
          type: string
        street:
          type: string
        zip:
          type: string
    # TemplateVariable:
    #   title: TemplateVariable
    #   type: object
    #   properties:
    #     composite:
    #       type: boolean
    #     description:
    #       type: string
    #     name:
    #       type: string
    #     singular:
    #       type: boolean
    #     type:
    #       type: string
    #       enum:
    #         - COMPOSITE_PARAM
    #         - DOT
    #         - FRAGMENT
    #         - PATH_SEGMENT
    #         - PATH_STYLE_PARAMETER
    #         - PATH_VARIABLE
    #         - REQUEST_PARAM
    #         - REQUEST_PARAM_CONTINUED
    #         - RESERVED_STRING
    #         - SEGMENT
    #         - SIMPLE
    # UriTemplate:
    #   title: UriTemplate
    #   type: object
    #   properties:
    #     variableNames:
    #       type: array
    #       items:
    #         type: string
    #     variables:
    #       type: array
    #       items:
    #         $ref: '#/components/schemas/TemplateVariable'
    UserDTO:
      title: UserDTO
      type: object
      properties:
        address:
          $ref: '#/components/schemas/AddressDTO'
        created:
          type: string
          format: date
        deleted:
          type: boolean
        dob:
          type: string
          format: date
        email:
          type: string
        fname:
          type: string
        fullName:
          type: string
        id:
          type: integer
          format: int64
        imageURL:
          type: string
        lname:
          type: string
        phone:
          type: string
        role:
          type: string
        updated:
          type: string
          format: date
    UserDTOReq:
      title: UserDTOReq
      type: object
      properties:
        address:
          $ref: '#/components/schemas/AddressDTO'
        created:
          type: string
          format: date
        deleted:
          type: boolean
        dob:
          type: string
          format: date
        email:
          type: string
        fname:
          type: string
        id:
          type: integer
          format: int64
        imageURL:
          type: string
        lname:
          type: string
        phone:
          type: string
        role:
          type: string
        updated:
          type: string
          format: date
    UserDTORes:
      title: UserDTORes
      type: object
      properties:
        address:
          $ref: '#/components/schemas/AddressDTO'
        created:
          type: string
          format: date
        deleted:
          type: boolean
        dob:
          type: string
          format: date
        email:
          type: string
        fname:
          type: string
        fullName:
          type: string
        id:
          type: integer
          format: int64
        imageURL:
          type: string
        lname:
          type: string
        phone:
          type: string
        role:
          type: string
        updated:
          type: string
          format: date
...
  securitySchemes:
    Authorization:
      type: apiKey
      name: Authorization
      in: header